+++ src/api/client.ts
@@ -60,9 +60,9 @@ export class Function {
     public constructor (config?: FunctionConfig) {
         this.client = new GraphClient(config ?? { });
         this.users = new UserService(this.client);
-        this.predictors = new PredictorService(this.client);
-        this.predictions = new PredictionService(this.client);
-        this.environmentVariables = new EnvironmentVariableService(this.client);
         this.storage = new StorageService(this.client);
+        this.predictors = new PredictorService(this.client);
+        this.predictions = new PredictionService(this.client, this.storage);
+        this.environmentVariables = new EnvironmentVariableService(this.client);
     }
 }