+++ test/prediction.ts
@@ -22,7 +22,21 @@ class PredictionTest {
     @mocha.test
     async "Should create a cloud prediction" () {
         const tag = "@natml/identity";
-        const inputs = { name: "Yusuf", age: 24 };
+        const inputs = {
+            name: "Yusuf",
+            age: 24,
+            ratio: 3.14159,
+            option: false,
+            choices: [true, false, true, true, false],
+            numbers: [12, 38, 4, 102, 99],
+            fractions: [4.2, 9.3, 1.3],
+            names: ["Drake", "Josh", "Andy"],
+            properties: {
+                length: 12,
+                height: 44,
+                width: 3.9
+            }
+        };
         const prediction = await this.fxn.predictions.create({ tag, inputs }) as CloudPrediction;
         const results = prediction.results;
         expect(results?.[0]).to.equal(`Yusuf`)